pythondaq/pyproject.toml
[tool.poetry]
name = "pythondaq"
version = "0.1.0"
description = ""
authors = ["JBusink <94122040+JBusink@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.1.3"
pyvisa-py = "^0.5.3"
pyserial = "^3.5"
pandas = "^1.5.1"
matplotlib = "^3.6.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
cli = "pythondaq.view.cli:cmd_group"